+++ integrations/grafana/integration.definition.ts
@@ -1,9 +1,8 @@
 import { IntegrationDefinition } from '@botpress/sdk'
 import { configuration, states, events, actions } from './definitions'
-import { integrationName } from './package.json'
 
 export default new IntegrationDefinition({
-  name: integrationName,
+  name: 'grafana',
   version: '0.1.0',
   readme: 'hub.md',
   icon: 'icon.svg',
