+++ integrations/grafana/src/gen/types/GrafanaDashboard.ts
