+++ integrations/grafana/tsconfig.json
@@ -1,27 +1,8 @@
 {
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "lib": ["es2022", "dom"],
-    "module": "preserve",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "bundler",
-    "allowUnusedLabels": false,
-    "allowUnreachableCode": false,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitOverride": true,
-    "noImplicitReturns": true,
-    "noUncheckedIndexedAccess": true,
-    "noUnusedParameters": true,
-    "target": "es2017",
     "paths": { "*": ["./*"] },
-    "outDir": "dist",
-    "checkJs": false,
-    "exactOptionalPropertyTypes": false,
-    "resolveJsonModule": true,
-    "noPropertyAccessFromIndexSignature": false,
-    "noUnusedLocals": false
+    "outDir": "dist"
   },
-  "include": [".botpress/**/*", "src/**/*", "*.ts", "*.json"]
+  "include": [".botpress/**/*", "definitions/**/*", "src/**/*", "*.ts"]
 }
